Minimum RAM requirement
When running `docker-compose up` it will take a very long time (or run out of memory) unless minimum RAM is set to 8GB.
This was tested using a MacBook Pro.
Docker Desktop for Mac by default gives 2GB to Docker. I had to change Docker Desktop to give 8GB and then it worked as normal.