Add LDAP Timeout Settings to Documentation
From the hidden configuration settings for LDAP authentication, the timeout settings should be made visible.

A recent run-in with an AD server that makes use of 2FA required a deep dive into the codebase.  The timeout settings were found, tested and works as expected, resulting in an acceptable timeframe to complete an AD login backed by a 2FA process.

The connection timeout affects the amount of time that is allowed to complete the 2FA process (at least in my case).  It's suggested to leave the response timeout at the default of 0 and adjust the connection timeout according to needs.

Reference: https://github.com/opendistro-for-elasticsearch/security/issues/211